requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Document processing
pdfminer.six==20231228
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0
chardet==5.2.0

# Embeddings and Vector DB
sentence-transformers==2.3.1
chromadb==0.4.22
rank-bm25==0.2.2

# LLM Integration
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# ML Forecasting
prophet==1.1.5
xgboost==2.0.3
scikit-learn==1.4.0

# Visualization
plotly==5.18.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.5.2
tenacity==8.2.3
